COG-1118: UserService interface is removed, UserServiceImpl renamed to UserService
